Finalisation du merge du coté Client + ajout d'un boolean dans VirtualCRM pour ne faire des requêtes vers Salesforce que quand aucun merge n'a été fait

+++ virtualcrm/src/main/java/fr/univangers/service/implementations/VirtualCRMServiceImpl.java
@@ -4,6 +4,7 @@ import fr.univangers.clients.GeoLocalisationServiceClient;
 import fr.univangers.clients.InternalCRMClient;
 import fr.univangers.clients.SalesforceCRMClient;
 import fr.univangers.internalcrm.thrift.*;
+import fr.univangers.merge.LeadMerger;
 import fr.univangers.model.GeographicPointDto;
 import fr.univangers.model.VirtualLeadDto;
 import fr.univangers.service.interfaces.IVirtualCRMService;
@@ -20,6 +21,9 @@ public class VirtualCRMServiceImpl implements IVirtualCRMService {
     private final InternalCRMClient internalCRMClient;
     
     private final GeoLocalisationServiceClient geoLocalisationServiceClient;
+
+    // True if a merge has been done
+    private boolean merged = false;
 
     public VirtualCRMServiceImpl(SalesforceCRMClient salesforceCRMClient,
                                  InternalCRMClient internalCRMClient,
@@ -42,10 +46,15 @@ public class VirtualCRMServiceImpl implements IVirtualCRMService {
      */
     @Override
     public List<VirtualLeadDto> findLeads(double lowAnnualRevenue, double highAnnualRevenue, String state) throws InvalidRevenueRangeException, TException, IOException {
-        List<VirtualLeadDto> leadsList = VirtualLeadConverter.mergeInternalSalesforceLeads(
-                salesforceCRMClient.findLeads(lowAnnualRevenue, highAnnualRevenue, state),
-                internalCRMClient.findLeads(lowAnnualRevenue, highAnnualRevenue, state)
-        );
+        List<VirtualLeadDto> leadsList = null;
+        if(!isMerged()) {
+            leadsList = VirtualLeadConverter.mergeInternalSalesforceLeads(
+                    salesforceCRMClient.findLeads(lowAnnualRevenue, highAnnualRevenue, state),
+                    internalCRMClient.findLeads(lowAnnualRevenue, highAnnualRevenue, state)
+            );
+        } else {
+            leadsList = internalCRMClient.findLeads(lowAnnualRevenue, highAnnualRevenue, state);
+        }
 
         for (VirtualLeadDto lead : leadsList) {
             Optional<GeographicPointDto> geographicPoint = geoLocalisationServiceClient.lookup(lead);
@@ -66,10 +75,16 @@ public class VirtualCRMServiceImpl implements IVirtualCRMService {
      */
     @Override
     public List<VirtualLeadDto> findLeadsByDate(long startDate, long endDate) throws InvalidDateException, TException, IOException {
-        List<VirtualLeadDto> leadsList = VirtualLeadConverter.mergeInternalSalesforceLeads(
-                salesforceCRMClient.findLeadsByDate(startDate, endDate),
-                internalCRMClient.findLeadsByDate(startDate, endDate)
-        );
+        List<VirtualLeadDto> leadsList = null;
+
+        if(!isMerged()){
+            leadsList = VirtualLeadConverter.mergeInternalSalesforceLeads(
+                    salesforceCRMClient.findLeadsByDate(startDate, endDate),
+                    internalCRMClient.findLeadsByDate(startDate, endDate)
+            );
+        } else {
+            leadsList = internalCRMClient.findLeadsByDate(startDate, endDate);
+        }
 
         for (VirtualLeadDto lead : leadsList) {
             Optional<GeographicPointDto> geographicPoint = geoLocalisationServiceClient.lookup(lead);
@@ -97,7 +112,9 @@ public class VirtualCRMServiceImpl implements IVirtualCRMService {
             lead = internalCRMClient.getLeadById(internalId);
         } catch (NumberFormatException e) {
             // Cas Salesforce : ID non numérique
-            lead = salesforceCRMClient.getLeadById(id);
+            if(!isMerged()) {
+                lead = salesforceCRMClient.getLeadById(id);
+            }
         }
 
         if (lead == null) {
@@ -140,12 +157,14 @@ public class VirtualCRMServiceImpl implements IVirtualCRMService {
             // pas trouvé dans InternalCRM donc on continue vers Salesforce
         }
 
-        // Essai Salesforce
-        try {
-            salesforceCRMClient.deleteLead(id);
-            return true;
-        } catch (LeadNotFoundException e) {
-            // pas trouvé dans Salesforce donc on laisse tomber
+        if(!isMerged()) {
+            // Essai Salesforce
+            try {
+                salesforceCRMClient.deleteLead(id);
+                return true;
+            } catch (LeadNotFoundException e) {
+                // pas trouvé dans Salesforce donc on laisse tomber
+            }
         }
 
         // Si aucun des deux n'a réussi
@@ -182,10 +201,15 @@ public class VirtualCRMServiceImpl implements IVirtualCRMService {
      */
     @Override
     public List<VirtualLeadDto> getLeads() throws TException, IOException {
-        List<VirtualLeadDto> leadsList = VirtualLeadConverter.mergeInternalSalesforceLeads(
-                salesforceCRMClient.getLeads(),
-                internalCRMClient.getLeads()
-        );
+        List<VirtualLeadDto> leadsList = null;
+        if(!isMerged()) {
+            leadsList = VirtualLeadConverter.mergeInternalSalesforceLeads(
+                    salesforceCRMClient.getLeads(),
+                    internalCRMClient.getLeads()
+            );
+        } else {
+            leadsList = internalCRMClient.getLeads();
+        }
 
         for (VirtualLeadDto lead : leadsList) {
             if(!lead.getCountry().isEmpty())
@@ -206,7 +230,28 @@ public class VirtualCRMServiceImpl implements IVirtualCRMService {
      */
     @Override
     public int countLeads() throws TException {
-        int countLeads = salesforceCRMClient.countLeads() + internalCRMClient.countLeads();
+        int countLeads = 0;
+        if(!isMerged()) {
+            countLeads = salesforceCRMClient.countLeads() + internalCRMClient.countLeads();
+        } else {
+            countLeads = internalCRMClient.countLeads();
+        }
         return countLeads;
+    }
+
+    @Override
+    public String merge() throws TException, IOException {
+        LeadMerger merger = new LeadMerger();
+        merger.merge();
+        setMerged(true);
+        return "Merge terminé";
+    }
+
+    public boolean isMerged() {
+        return merged;
+    }
+
+    public void setMerged(boolean merged) {
+        this.merged = merged;
     }
 }
